# frontend/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "A JWT signed with the secret of the server, sent as `Authorization: Bearer <token>` (`Authorization: Token <token>`, which InfluxDB clients send, works too). Each operation lists the scope it needs: `read`, `write`, `delete` or `admin`. Make one with `sensapp generate-token`, or with `POST /api/v1/admin/tokens` and an `admin` token (the Credentials tab of the web UI does). With authentication disabled (`SENSAPP_AUTH_DISABLED`) no token is needed.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Sensor data ingestion, storage, and retrieval platform",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "sensapp",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/query": {
      "get": {
        "description": "Parses a PromQL expression and returns matching time series data.\nOnly simple selectors (VectorSelector and MatrixSelector) are supported.\n\n# Examples\n\n- Simple metric: `GET /api/v1/query?query=my_metric`\n- With labels: `GET /api/v1/query?query=my_metric{env=\"prod\"}`\n- Range query: `GET /api/v1/query?query=my_metric[5m]`\n- With format: `GET /api/v1/query?query=my_metric&format=csv`",
        "operationId": "simple_promql_query",
        "parameters": [
          {
            "description": "PromQL query string (e.g., 'my_metric{label=\"value\"}' or 'my_metric[5m]')",
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Output format: senml (default), csv, jsonl, or arrow",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Bucket width for aggregation queries, using Prometheus duration syntax (e.g., '5m'). Without it, the whole window is one bucket",
            "in": "query",
            "name": "step",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Query results in requested format"
          },
          "400": {
            "description": "Invalid or unsupported PromQL query"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "The storage backend is unavailable: retry later"
          },
          "504": {
            "description": "The request took longer than SENSAPP_HTTP_SERVER_TIMEOUT_SECONDS, usually because the storage backend hangs"
          }
        },
        "security": [
          {
            "bearer": [
              "read"
            ]
          }
        ],
        "summary": "Simple PromQL query endpoint.",
        "tags": [
          "SensApp"
        ]
      }
    },
...
  },
...
}